refactor(runner): move kernel warmup into the shared runner lifecycle (warmup())
Kernel warmup / flashinfer autotune is part of the runner lifecycle, not a
standalone ModelRunner init step. Now that EagerRunner exists as a BaseRunner
sibling, warmup() lives on the shared BaseRunner so BOTH the cuda-graph runners
AND the eager runner inherit it — and the eager-mode warmup is no longer a
special case in init_backends.

  - BaseRunner.warmup() holds the orchestration directly (run-once via a
    `_kernel_warmed_up` flag on the shared ModelRunner; device guard; flashinfer
    allreduce pre-init; flashinfer autotune; PP-DeepGEMM warmup).
  - Decode/PrefillCudaGraphRunner warm up from capture() (autotune right before
    the graph it tunes), exactly as before.
  - EagerRunner has no capture step, so it warms up itself: warmup() on its
    first execute() (cheap no-op thereafter via the run-once flag).
  - DELETE the eager kernel_warmup() fallback from init_backends entirely; eager
    mode warms up purely via EagerRunner.warmup() on first execute.

Move the autotune / dummy-run machinery off ModelRunner onto BaseRunner so
model_runner.py owns NONE of it; base_runner.py owns it all. Behavior-identical
— the moved methods read shared ModelRunner state through self.model_runner:

  - _should_run_flashinfer_autotune, _flashinfer_autotune,
    _flashinfer_autotune_cache_path, _dummy_run,
    _pre_initialize_flashinfer_allreduce_workspace → BaseRunner methods.
  - _allocate_decode_buffers → module-level function in base_runner.py (moved
    out of decode_cuda_graph_runner.py); _dummy_run self-allocates via it.
  - init_backends no longer calls _pre_initialize_flashinfer_allreduce_workspace;
    warmup() calls it FIRST (after the run-once guard, before autotune), so it
    still runs before cuda-graph capture (capture() calls warmup() before the
    graph_capture() context). The method name _flashinfer_autotune is unchanged,
    so pynccl_allocator's traceback string-check still matches.
  - pp_parallel_deep_gemm_warmup now takes the RUNNER (BaseRunner.warmup() passes
    self) and calls runner._dummy_run(...); ModelRunner state via
    runner.model_runner.

Drops the now-dead imports from model_runner.py (ceil_align, require_mlp_sync,
DpPaddingMode, set_dp_buffer_len, set_is_extend_in_batch, get_attention_tp_size,
require_mlp_tp_gather, require_attn_tp_gather, CaptureHiddenMode,
create_dummy_verify_input, set_torch_compile_config, empty_context, hashlib,
Path, _allocate_decode_buffers) and from decode_cuda_graph_runner.py
(NgramEmbeddingInfo, envs).

Test mock: MockModelRunner sets _kernel_warmed_up=True — it never calls
init_backends and has no real kernels, so warmup is a no-op for the runner-mode
attention tests that drive capture() / execute() directly.
