+++ python/sglang/test/kits/attention_unittest/attention_methods/dual_chunk_attention.py
@@ -321,6 +321,7 @@ class DualChunkMockModelRunner(ModelRunner):
         self.page_size = case.page_size
         self.model_config = model_config
         self.tp_size = 1
+        self._kernel_warmed_up = True
         self.dp_size = 1
         self.pp_size = 1
         self.server_args = make_mock_server_args(
