+++ python/sglang/test/kits/attention_unittest/attention_methods/gdn_attention.py
@@ -304,6 +304,7 @@ class MockGDNModelRunner(ModelRunner):
         self.sliding_window_size = None
         self.use_mla_backend = False
         self.is_draft_worker = False
+        self._kernel_warmed_up = True
 
     @property
     def hybrid_gdn_config(self):
