+++ python/sglang/test/kits/attention_unittest/attention_methods/kda_attention.py
@@ -310,6 +310,7 @@ class MockKDAModelRunner(ModelRunner):
         self.sliding_window_size = None
         self.use_mla_backend = False
         self.is_draft_worker = False
+        self._kernel_warmed_up = True
 
     @property
     def hybrid_gdn_config(self):
