+++ python/sglang/test/kits/attention_unittest/attention_methods/mamba2_attention.py
@@ -454,6 +454,7 @@ class MockMamba2ModelRunner(ModelRunner):
         self.sliding_window_size = None
         self.use_mla_backend = False
         self.is_draft_worker = False
+        self._kernel_warmed_up = True
 
     @property
     def hybrid_gdn_config(self):
